修复 Apple sysctl 头包含顺序；x509 跳过证书后清除挂起错误；base64 反查表窄化转换；同步单头

+++ src/codec/base64.c
@@ -135,7 +135,7 @@ static bool __xrtBase64Prepare(
 
 		if ( !bCustom ) {
 			if ( pReverse != NULL ) {
-				pReverse[iCharacter] = (int16)i;
+				pReverse[iCharacter] = (int8)i;
 			}
 			continue;
 		}
